+++ tests/page.test.js
@@ -60,7 +60,7 @@ test('statoil', t => {
   t.equal(doc.categories().length, 4, 'cat-length');
   t.equal(doc.images().length, 1, 'img-length');
   t.equal(doc.images(0).file(), 'Fil:Statoil-Estonia.jpg', '');
-  t.equal(doc.images(0).url(), 'https://upload.wikimedia.org/wikipedia/commons/8/87/Statoil-Estonia.jpg', t);
+  t.equal(doc.images(0).url(), 'https://wikipedia.org/wiki/Special:Redirect/file/Statoil-Estonia.jpg', t);
   t.end();
 });
 
@@ -73,7 +73,7 @@ test('raith rovers', t => {
   t.equal(doc.images(1).file(), 'File:Stark\'s Park - geograph.org.uk - 204446.jpg', 'img-file');
   t.equal(
     doc.images(1).url(),
-    'https://upload.wikimedia.org/wikipedia/commons/3/38/Stark\'s_Park_-_geograph.org.uk_-_204446.jpg',
+    `https://wikipedia.org/wiki/Special:Redirect/file/Stark's_Park_-_geograph.org.uk_-_204446.jpg`,
     'image-url'
   );
   t.end();